Register "Page Setup...", "Print..."; break FileMenus -> Printing...
Also FileMenus -> TrackPanel
