+++ src/Printing.cpp
@@ -14,9 +14,6 @@
 *//*******************************************************************/
 
 
-
-#include "Printing.h"
-
 #include <wx/defs.h>
 #include <wx/dc.h>
 #include <wx/intl.h>
@@ -35,6 +32,13 @@
 
 #include "tracks/ui/TrackView.h"
 
+#include "commands/CommandContext.h"
+#include "commands/CommandManager.h"
+#include "CommonCommandFlags.h"
+#include "Project.h"
+#include "TrackPanel.h"
+
+namespace {
 // Globals, so that we remember settings from session to session
 wxPrintData &gPrintData()
 {
@@ -190,4 +194,40 @@ void HandlePrint(
    else {
       gPrintData() = printer.GetPrintDialogData().GetPrintData();
    }
+}
+
+void OnPageSetup(const CommandContext &context)
+{
+   auto &project = context.project;
+   auto &window = GetProjectFrame( project );
+   HandlePageSetup(&window);
+}
+
+void OnPrint(const CommandContext &context)
+{
+   auto &project = context.project;
+   auto name = project.GetProjectName();
+   auto &tracks = TrackList::Get( project );
+   auto &window = GetProjectFrame( project );
+   HandlePrint(&window, name, &tracks, TrackPanel::Get( project ));
+}
+
+using namespace MenuTable;
+BaseItemSharedPtr PrintingItems()
+{
+   static BaseItemSharedPtr items{
+   Section( "Print",
+      Command( wxT("PageSetup"), XXO("Pa&ge Setup..."), OnPageSetup,
+         AudioIONotBusyFlag() | TracksExistFlag() ),
+      /* i18n-hint: (verb) It's item on a menu. */
+      Command( wxT("Print"), XXO("&Print..."), OnPrint,
+         AudioIONotBusyFlag() | TracksExistFlag() )
+   ) };
+   return items;
+}
+
+AttachedItem sAttachment{ { "File", { OrderingHint::Before, "Exit" } },
+   Shared( PrintingItems() )
+};
+
 }
