+++ src/menus/FileMenus.cpp
@@ -4,7 +4,6 @@
 #include "../NoteTrack.h"
 #include "PluginManager.h"
 #include "Prefs.h"
-#include "../Printing.h"
 #include "Project.h"
 #include "../ProjectFileIO.h"
 #include "../ProjectFileManager.h"
@@ -14,7 +13,6 @@
 #include "../ProjectWindow.h"
 #include "../SelectFile.h"
 #include "../SelectUtilities.h"
-#include "../TrackPanel.h"
 #include "UndoManager.h"
 #include "ViewInfo.h"
 #include "../WaveTrack.h"
@@ -533,22 +531,6 @@ void OnImportRaw(const CommandContext &context)
    DoImport(context, true);
 }
 
-void OnPageSetup(const CommandContext &context)
-{
-   auto &project = context.project;
-   auto &window = GetProjectFrame( project );
-   HandlePageSetup(&window);
-}
-
-void OnPrint(const CommandContext &context)
-{
-   auto &project = context.project;
-   auto name = project.GetProjectName();
-   auto &tracks = TrackList::Get( project );
-   auto &window = GetProjectFrame( project );
-   HandlePrint(&window, name, &tracks, TrackPanel::Get( project ));
-}
-
 void OnExit(const CommandContext &WXUNUSED(context) )
 {
    // Simulate the application Exit menu item
@@ -702,14 +684,6 @@ BaseItemSharedPtr FileMenu()
             Command( wxT("ImportRaw"), XXO("&Raw Data..."), OnImportRaw,
                AudioIONotBusyFlag() )
          )
-      ),
-
-      Section( "Print",
-         Command( wxT("PageSetup"), XXO("Pa&ge Setup..."), OnPageSetup,
-            AudioIONotBusyFlag() | TracksExistFlag() ),
-         /* i18n-hint: (verb) It's item on a menu. */
-         Command( wxT("Print"), XXO("&Print..."), OnPrint,
-            AudioIONotBusyFlag() | TracksExistFlag() )
       ),
 
       Section( "Exit",
